Equb/requirements.txt
aioredis==2.0.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
Django==5.0.4
django-compat==1.0.15
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-guardian==2.4.0
django-rest-passwordreset==1.4.1
django4-background-tasks==1.2.9
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
hiredis==2.3.2
hyperlink==21.0.0
idna==3.7
importlib_metadata==7.1.0
incremental==22.10.0
Markdown==3.6
msgpack==1.0.8
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyHamcrest==2.1.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pytz==2024.1
redis==5.0.4
service-identity==24.1.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
zipp==3.18.1
zope.interface==6.3
